requirements.txt
pytest>=6.2.3
requests>=2.8.40
jsonschema>=3.2.0
selenium>=3.141.0
allure-pytest>=2.8.40
beautifulsoup4>=4.9.3